Add swapchain frame buffers
Will need a future refactor, the dependency tree wasn't fully understood
when the files were broken out, and now things have to be passed back
and forth in some places such as the VkRenderPass. Refactoring should
be left as a TODO until after an image is on screen and the full
dependency tree is fully understood.