Refactor logic to move control from device to main
Too command buffers, swapchains, etc were all handled in device since
they all used VkDevice as a dependency. Refactor now gets the logical
and physical device from device.cpp and handles the program flow in there.

The idea is to let device.cpp handle the lifecycle of the logical device
only, but not the lifecycle of other vulkan components.