Add component
The idea is to store a set of components for each entity (ECS) that
will clean themselves up. This component class allows the user to
generically wrap any struct into the component. In order to cast back
the plan is to generate some c++ code for each struct in the src/components
folder and cast based on the type_info