Add builder for Engine and support for validatoin layers
Builder may be useful in future as more and more properties begin
to accumulate in the Engine struct, and this allows consumers to
easily pass in optional values.

event_loop has to be optional in order to swap it, since it EventLoop<()>
has no default value. This also allows swap to give a reasonable value back,
if the engine tries to access it again it will see the None value.